custom hook -> AppDownloadLink
